backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic>=2.10,<3
pydantic-settings>=2.7,<3
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
pandas==2.2.3
openpyxl==3.1.5
xlrd>=2.0.1
xlsxwriter>=3.1.0
fpdf2>=2.7.0
rembg>=2.0.50
openai>=1.0.0
Pillow>=10.0.0

